requirements/base.txt
# Base
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.3.1  # https://github.com/python-pillow/Pillow
psycopg2==2.9.1  # https://github.com/psycopg/psycopg2
PyJWT==1.6.4

# Django
django==3.2.5

# Django REST Framework
djangorestframework-jwt==1.11.0
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-filter==2.4.0

# Cors
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi

# Static files
whitenoise==5.3.0

# Celery
redis==3.5.3  # https://github.com/andymccurdy/redis-py
django-redis==5.0.0  # https://github.com/jazzband/django-redis
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat

# django-allauth==0.45.0 # https://github.com/pennersr/django-allauth
